Stop using symbolic link
